town_clerks/requirements.txt
Django>=4.2
python-dotenv
mssql-django
django-auth-adfs
pandas
pyodbc
openpyxl
